pyproject.toml
[project]
name = "goit-pythonweb-hw-12"
version = "0.1.0"
description = "REST API-застосунок"
authors = [
    {name = "aleksandr", email = "aleksandr.rebenok@ukr.net"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"  

dependencies = [
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "alembic (>=1.16.1,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "pydantic[email] (>=2.11.5,<3.0.0)",
    "greenlet (>=3.2.2,<4.0.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "fastapi-mail (>=1.5.0,<2.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "cloudinary (>=1.44.0,<2.0.0)",
    "slowapi (>=0.1.9,<0.2.0)",
    "libgravatar (>=1.0.4,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "httpx (>=0.28.1,<0.29.0)",
    "asgi-lifespan (>=2.1.0,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "redis (>=6.2.0,<7.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
sphinx = "5.3.0"
sphinx-autodoc-typehints = "1.22"
pytest = "^8.4.0"
pytest-asyncio = "^1.0.0"
pytest-cov = "^6.2.1"
aiosqlite = "^0.21.0"
httpx = "^0.28.1"

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::DeprecationWarning"
]